Feature/abt 5320 base cypress integration (#14)
* Integrated Cypress

This commit contains the salesforce-cypress repo. Going forward, all Cypress tests for salesforce will be created within the Salesforce repo

* Corrected cypress example doc

* Updated Defect Fix

Updated data based on scratch org

* Scratch org fix

* Modified login command

This fixes an issue that was occurring when running all the tests back to back.

* scratch org fixes for case clicking

* Removed complaint tests

To prep the Cypress integration, this branch contains only the base Cypress integration as well as some common functions. The complaints tests have been removed and will be implemented in another feature branch.

* Updated readme and removed example fixture

Co-authored-by: Mukesh Kumar Jha <mukeshkumar.jha@anz.com>